Refactor Nikon protocol handling.
Split the monolithic Nikon protocol into the 2 known instances: remote
and smart.
This keeps the protocols self-contained and simpler to maintain going
forward.
Nikon appears to retain the higher level handshake, so NikonBase drives
that and delegates to the specialisation when needed.
